requirements.txt
streamlit
pandas
plotly
numpy
Pillow
pandas
requests
tweepy
numpy
scipy
seaborn
streamlit
jupyterlab
bokeh
plotly
kaleido